Ensure the update (if there is one) shows up right away when the user clicks "Check for updates". Prior to this change they would need to reload the page to see the update.
Cause: Originally the update check routine was attached to the admin_notices hook. However, the plugin list is populated long before that action. 

Fixed by running the update check earlier in the request process and redirecting back to plugins.php to display the success message. As a side-effect, this will also prevent unwanted extra checks when the user reloads the page.